luci-mod-network: wireless.js: allow changing the 60 GHz radio's frequency
The text "Operating frequency" was still on the form, but there was no
drop-down next to it. Also, it was only possible to edit the 60 GHz
radio's settings once: After that, instead of a settings dialog, an
error dialog would appear that said "TypeError: chan.options[0] is
undefined".

"mhz >= 58329" appears to have been a typo. The TP-Link AD7200 reports a
channel at 58320 MHz, exactly where the 802.11 specification says it
should be.

Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>